Game Over screen Fixed
Fixed colors, positions and behavior of elements once the game is over ( due to a collision ) . Game Over sign is now bigger and different color. Implemented 2 new methods to hide or show the obstacles. When the doggie collides with any of the pipes ( or ground or sky ) it hides the obstacles and vice versa the obstacles  are reset once the Retry button is clicked.